feat(core): add watermarkScope marker; extract health types to sibling (#2151)
Round-3 review + #2149-merge follow-up on the extraction-liveness verdict.

watermarkScope marker (review): the /health `extraction` block, the
`extraction_liveness` doctor check, and `remnic stats` now carry a
`watermarkScope` field ("root-store" today). A monitor can alert on
`watermarkScope != "aggregate"` while namespaces are enabled instead of
trusting a verdict that may be wrong in either direction under multiple
namespaces. It flips to "aggregate" when #2159 lands, so the limitation
leaves the payload by construction, not by editing prose.

Type-cluster extraction (ratchet remedy after the #2149 merge): moved
EngramAccessHealthResponse, EngramAccessQmdCollectionState, and
EngramAccessQmdHealthResponse verbatim from access-service.ts into a new
sibling access-health-types.ts, re-exported from access-service.ts so
existing importers resolve unchanged. Rationale: #2151 (extraction
liveness) and #2149 (corpus watermarks) were each engineered to land
exactly at access-service.ts's grandfathered 5911-line ceiling; the
mandatory merge sums them to 5916 - neither PR was individually over. The
five added lines are irreducible wiring (import, interface field, instance
field, resolve call, return property) whose substantive logic already
lives in sibling modules. The ratchet ratchets downward only and refuses
to raise a grandfathered ceiling, so the sanctioned remedy is a sibling
extraction; it also leaves access-service.ts genuinely smaller (5883), not
parked at a new high-water mark. Type-only relocation, zero logic changed.

Part of #2151
