+++ packages/remnic-core/src/access-http.test.ts
@@ -1649,6 +1649,7 @@ test("HTTP externalRequestHandler runs pre-auth, can end responses, and falls th
         oldestBufferedTurnAgeMs: null,
         degraded: false,
         degradedReason: null,
+        watermarkScope: "root-store",
       },
       corpus: [],
     }),
@@ -1773,6 +1774,7 @@ test("HTTP authTokenEntriesGetter is authoritative: scope policy binds connector
         oldestBufferedTurnAgeMs: null,
         degraded: false,
         degradedReason: null,
+        watermarkScope: "root-store",
       },
       corpus: [],
     }),
