+++ packages/remnic-server/src/oauth.test.ts
@@ -85,6 +85,7 @@ const serviceStub = {
       oldestBufferedTurnAgeMs: null,
       degraded: false,
       degradedReason: null,
+      watermarkScope: "root-store",
     },
     corpus: [],
   }),
